Check how many args are being passed (#15)
* check how many args

Check to make sure that at least 2 args are being passed to mosbot,
this will avoid pulling bugs/docs with the id set to 0.

* Additional check

Realized that if '-o' or '-c' is passed in it would still be
possible for the id to be empty. So added an additional check
to see if the id is 0 where the type is checked.