Proposal for #13 Grab Document title from public page (#16)
* check how many args

Check to make sure that at least 2 args are being passed to mosbot,
this will avoid pulling bugs/docs with the id set to 0.

* Additional check

Realized that if '-o' or '-c' is passed in it would still be
possible for the id to be empty. So added an additional check
to see if the id is 0 where the type is checked.

* Proposal for #13 Grab Document title from public page

Added new function 'generate_public_url', I'm not sure that it handles
all cases though.

Added new function 'get_title', uses regex to get the page title.

Additional code to use these 2 functions and print the page title.

* Revert "Additional check"

This reverts commit cde04dcba078ce7af0f618a1bbbeecaded62ad47.
Shouldn't have been in this branch yet.

* Revert "check how many args"

This reverts commit 201d2930c9efe3d7cb0304c7796804fd271a573a.
Also shouldn't have been in this branch yet.

* Add error handling

Added some error handling to the 'get_title' function, it's very
basic but should prevent dumping errors should they arise.

Also adding title empty check to output to only print if there's a
title available.